refactor user dtos(change Optional to '|')
